крупный рефакторинг (TODO: переделать на QConcurrent)

+++ theodolitemeasure.cpp
@@ -5,12 +5,12 @@ TheodoliteMeasure::TheodoliteMeasure()
      measure_tuple=std::make_tuple(0, 0, 0);
 }
 
-TheodoliteMeasure::TheodoliteMeasure(const double& hz_angle, const double& v_angle, const quint16 circle)
+TheodoliteMeasure::TheodoliteMeasure(const double& hz_angle, const double& v_angle, const qint32 circle)
 {
     measure_tuple=std::make_tuple(hz_angle, v_angle,circle+1);
 }
 
-void TheodoliteMeasure::setMeasureData(const double& hz_angle, const double& v_angle, const quint16 circle)
+void TheodoliteMeasure::setMeasureData(const double& hz_angle, const double& v_angle, const qint32 circle)
 {
     measure_tuple=std::make_tuple(hz_angle, v_angle,circle+1);
 }
@@ -24,7 +24,7 @@ double TheodoliteMeasure::getVAngle() const
 {
     return std::get<1> (measure_tuple);
 }
-quint16 TheodoliteMeasure::getCircle() const
+qint32 TheodoliteMeasure::getCircle() const
 {
     return std::get<2>(measure_tuple);
 }
